Package rename - Setting up P.2
I should have added the team to the packages and not just my name, but I'm that egocentric so now I have to change it.
pi stands for Plan Infalible (your guess, it means Infalible Plan... Duh!) and it's the group of Efrik and myself. I do the code, he does the art.